added fetcher
